feat(container): make base images configurable
With this commit, we use build args to declare the base image
of the appserver image (defaulting to openjdk:11-jre) and
declare the Solr version.

This replaces variable substitution by the Docker Maven Plugin.
Primary reason to do this was to enable reuse of the Dockerfiles
with Testcontainers, which does not support variable substitution.

Labels still have DMP variables, but this is not important for
the Testcontainers use case.